requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==2015.4.28
Django==1.8.2
django-allauth==0.20.0
django-appconf==1.0.1
django-bootstrap3==6.1.0
django-filter==0.10.0
django-geoposition==0.2.2
django-model-utils==2.2
facebook-sdk==0.4.0
functools32==3.2.3.post1
geopy==1.10.0
googlemaps==2.2
httplib2==0.9.1
ipython==3.1.0
Jinja2==2.7.3
jsonschema==2.5.1
Markdown==2.6.2
MarkupSafe==0.23
mistune==0.6
oauth2==1.5.211
oauthlib==0.7.2
psycopg2==2.6.1
ptyprocess==0.5
Pygments==2.0.2
python-openid==2.2.5
pytz==2015.4
pyzmq==14.6.0
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
terminado==0.5
tornado==4.2
